requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgi-redis==0.10.0
asgiref==0.13.0
autobahn==0.14.0
billiard==3.3.0.23
celery==3.1.23
channels==0.12.0
daphne==0.12.0
Django==1.9.6
django-audiofield==0.8.2
django-celery==3.1.17
django-extensions==1.6.6
django-filter==0.13.0
django-guardian==1.4.4
djangorestframework==3.3.3
drf-ujson==1.2
gunicorn==19.4.5
kombu==3.0.35
msgpack-python==0.4.7
oauthlib==1.1.1
Pillow==3.2.0
psycopg2==2.6.1
pytz==2016.4
redis==2.10.5
requests==2.10.0
requests-oauthlib==0.6.1
six==1.10.0
Twisted==16.1.1
txaio==2.5.1
ujson==1.35
zope.interface==4.1.3
git+https://github.com/khasanovbi/rest_channels.git
rollbar
